Use a web worker attempt 1 (with error)
