+++ staex/src/redux/action-creators.js
@@ -1,16 +1,6 @@
 import store from './store'
-import MCTS from '../mcts/'
-import Game from '../mcts/game'
 import { winnerSelector } from './selectors'
-
-function computerMove(state) {
-	setTimeout(() => {
-		const game = new Game({ initialState: state })
-		const mcts = new MCTS({ game, rounds: 100000, timeout: 60000 })
-		const move = mcts.getMove()
-		store.dispatch(move)
-	}, 10)
-}
+import { getComputerMove } from './computer'
 
 export const stack = payload => {
 	// Player move
@@ -19,7 +9,7 @@ export const stack = payload => {
 	if (winnerSelector(store.getState())) return
 
 	// Computer move
-	computerMove(store.getState())
+	getComputerMove(store.getState())
 }
 
 export const move = payload => {
@@ -29,5 +19,5 @@ export const move = payload => {
 	if (winnerSelector(store.getState())) return
 
 	// Computer move
-	computerMove(store.getState())
+	getComputerMove(store.getState())
 }
